requirements.txt
PyYAML==3.10
argparse==1.2.1
blessings==1.5.1
bzr==2.6.0
bzr+http://code.launchpad.net/~bcsaller/charmtester/bundletester#egg=bundletester
charm-tools==1.3.2
charmworldlib==0.4.1
coverage==3.7.1
juju-deployer==0.3.8
mock==1.0.1
nose==1.3.1
requests==2.3.0
wsgiref==0.1.2
